server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.0
spacy==3.7.2
textstat==0.7.3
requests==2.31.0
werkzeug==2.3.7